pyproject.toml
[tool.poetry]
name = "watermarking"
version = "0.1.0"
description = "Watermarking as Defense against Model Extraction Attacks on Deep Learning Models"
authors = ["Ridhima Garg <garg.ridhima72@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
# Core ML Frameworks
tensorflow-metal = "*"  # Metal GPU acceleration for macOS (Apple Silicon)
# tensorflow-addons - Optional, version 0.23.0 doesn't have wheels for Python 3.12
# tensorflow-addons = "*"
# tensorflow-datasets - Optional, depends on tensorflow-metadata which conflicts with TF 2.20
# tensorflow-datasets = "*"
tensorflow-estimator = "*"
# tensorflow-io-gcs-filesystem - Optional, not compatible with Python 3.12
# tensorflow-io-gcs-filesystem = "*"
# tensorflow-metadata - Optional, has protobuf conflict with TF 2.20 (requires protobuf <5, but TF 2.20 requires protobuf >=5.28.0)
# tensorflow-metadata = "*"
# tensorflow-model-optimization - Optional, has numpy conflict with TF 2.19 (requires numpy <2.0, but TF 2.19 requires numpy >=2.1.0)
# tensorflow-model-optimization = "*"
# PyTorch - Optional, only needed for DAWN method
# PyTorch 2.0.1 doesn't support Python 3.13, use newer version or install separately
# For Python 3.13, install: pip install torch torchvision torchaudio
# torch = "^2.0.0"
# torchvision = "^0.15.0"
# torchaudio = "^2.0.0"

# Adversarial Robustness
adversarial-robustness-toolbox = "*"

# Data Processing
numpy = "*"
pandas = "*"
scipy = "*"
scikit-learn = "*"
scikit-image = "*"
Pillow = "*"
opencv-python = "*"
opencv-python-headless = "*"
h5py = "*"
tables = "*"

# Visualization
matplotlib = "*"
seaborn = "*"
plotly = "*"
altair = "*"
kaleido = "*"

# Configuration & Utilities
mlconfig = "*"
PyYAML = "*"
oyaml = "*"
click = "*"
tqdm = "*"
rich = "*"

# Web Framework (for Streamlit app)
streamlit = "*"
Flask = "*"
starlette = "*"
sse-starlette = "*"

# Model Optimization & Conversion
onnx = "*"
onnxruntime = "*"
# tf2onnx - Optional git dependency, install separately if needed
# tf2onnx = {git = "https://github.com/onnx/tensorflow-onnx.git", rev = "25c977c2e27e3ff793a275ebd6eec444d5c9f914"}

# DAWN Package (git dependency) - Optional, only needed for DAWN method
# Uncomment and configure if you need DAWN functionality
# dawndynamicadversarialwatermarkingofneuralnetworks = {git = "https://github.com/ridhimagarg/MasterThesis.git", rev = "291c3a8f885a4d888e72be5ed5036604f9425d44", subdirectory = "model-extraction-defense/modeldefense/dawndynamicadversarialwatermarkingofneuralnetworks"}

# Computer Vision & Image Processing
imageio = "*"
PyWavelets = "*"
tifffile = "*"

# Deep Learning Utilities
tensorboard = "*"
tensorboard-data-server = "*"
tensorboard-plugin-wit = "*"
thop = "*"
# triton - Optional, PyTorch dependency, not compatible with Python 3.13
# triton = "*"

# Data Management
fiftyone = "*"
fiftyone-brain = "*"
fiftyone-db = "*"
voxel51-eta = "*"

# Utilities
bitstring = "*"
dacite = "*"
dill = "*"
cloudpickle = "*"
joblib = "*"
retrying = "*"
tenacity = "*"
validators = "*"
humanfriendly = "*"
coloredlogs = "*"

# Network & HTTP
requests = "*"
httpx = "*"
httpcore = "*"
urllib3 = "*"
aiofiles = "*"

# Database
SQLAlchemy = "*"
pymongo = "*"
mongoengine = "*"
motor = "*"

# Authentication & Security
google-auth = "*"
google-auth-oauthlib = "*"
PyJWT = "*"
pycryptodomex = "*"
requests-oauthlib = "*"

# Jupyter & Notebooks
ipython = "*"
ipykernel = "*"
ipywidgets = "*"
jupyter_client = "*"
jupyter_core = "*"
jupyterlab-widgets = "*"
notebook = "*"

# Development Tools
ipdb = "*"
debugpy = "*"
pytest = "*"
black = "*"
flake8 = "*"

# System & OS
psutil = "*"
GPUtil = "*"
py-cpuinfo = "*"

# Compression & Archives
py7zr = "*"
rarfile = "*"
pyzstd = "*"
Brotli = "*"
inflate64 = "*"
pybcj = "*"
pyppmd = "*"
multivolumefile = "*"

# Parsing & Serialization
jsonlines = "*"
xmltodict = "*"
msgpack = "*"
protobuf = "*"
flatbuffers = "*"

# Date & Time
python-dateutil = "*"
pytz = "*"
pytz-deprecation-shim = "*"
tzdata = "*"
tzlocal = "*"
# backports-zoneinfo - Not needed for Python 3.9+ (zoneinfo is built-in), and incompatible with Python 3.12
# backports-zoneinfo = "*"

# String Processing
ftfy = "*"
regex = "*"
chardet = "*"
charset-normalizer = "*"

# GraphQL
strawberry-graphql = "*"
graphql-core = "*"

# Other Dependencies
absl-py = "*"
alembic = "*"
anyio = "*"
argcomplete = "*"
array-record = "*"
astropy = "*"
asttokens = "*"
astunparse = "*"
attrs = "*"
backcall = "*"
blinker = "*"
blosc2 = "*"
boto3 = "*"
botocore = "*"
build = "*"
cachetools = "*"
certifi = "*"
comm = "*"
contourpy = "*"
cycler = "*"
Cython = "*"
data = "*"
databricks-cli = "*"
decorator = "*"
Deprecated = "*"
dm-tree = "*"
dnspython = "*"
docker = "*"
entrypoints = "*"
etils = "*"
eventlet = "*"
exceptiongroup = "*"
executing = "*"
filelock = "*"
fonttools = "*"
funcsigs = "*"
future = "*"
gast = "*"
gitdb = "*"
GitPython = "*"
glob2 = "*"
google-pasta = "*"
googleapis-common-protos = "*"
greenlet = "*"
grpcio = "*"
gunicorn = "*"
h11 = "*"
h2 = "*"
hpack = "*"
hypercorn = "*"
hyperframe = "*"
idna = "*"
importlib-metadata = "*"
importlib-resources = "*"
itsdangerous = "*"
jedi = "*"
Jinja2 = "*"
jmespath = "*"
jsonschema = "*"
jsonschema-specifications = "*"
Keras-Preprocessing = "*"
keras = "*"
kiwisolver = "*"
latex = "*"
lazy_loader = "*"
libclang = "*"
lit = "*"
Mako = "*"
Markdown = "*"
markdown-it-py = "*"
MarkupSafe = "*"
matplotlib-inline = "*"
mdurl = "*"
mpmath = "*"
nest-asyncio = "*"
networkx = "*"
numexpr = "*"
oauthlib = "*"
opt-einsum = "*"
packaging = "*"
parso = "*"
pexpect = "*"
pickleshare = "*"
pip-tools = "*"
pkgutil_resolve_name = "*"
platformdirs = "*"
pprintpp = "*"
priority = "*"
promise = "*"
prompt-toolkit = "*"
pyarrow = "*"
pyasn1 = "*"
pyasn1-modules = "*"
pycocotools = "*"
pydeck = "*"
pydot = "*"
pyerfa = "*"
Pygments = "*"
Pympler = "*"
pyparsing = "*"
pyproject_hooks = "*"
pyzmq = "*"
querystring-parser = "*"
referencing = "*"
rpds-py = "*"
rsa = "*"
s3transfer = "*"
sentry-sdk = "*"
shutilwhich = "*"
six = "*"
smmap = "*"
sniffio = "*"
sortedcontainers = "*"
sqlparse = "*"
sseclient-py = "*"
stack-data = "*"
sympy = "*"
tabulate = "*"
tempdir = "*"
termcolor = "*"
texttable = "*"
threadpoolctl = "*"
toml = "*"
tomli = "*"
toolz = "*"
tornado = "*"
traitlets = "*"
typeguard = "*"
typing_extensions = "*"
ultralytics = "*"
universal-analytics-python3 = "*"
watchdog = "*"
wcwidth = "*"
websocket-client = "*"
Werkzeug = "*"
widgetsnbextension = "*"
wrapt = "*"
wsproto = "*"
zipp = "*"
tensorflow = "2.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
black = "*"
flake8 = "*"
mypy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

